shared/pyproject.toml
[project]
name = "network-chan-shared"
version = "0.1.0"
description = "Shared models, settings, utilities, and base ML components for Network-Chan (Appliance + Assistant)"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.12"
license = { text = "MIT" }  # Update if you choose a different license
authors = [{ name = "Dave", email = "your-email@example.com" }]
keywords = ["aiops", "sdn", "mlops", "networking"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    # Core Config and Utils
    "pydantic>=2.12.5",
    "pydantic-settings>=2.3",
    "pyyaml>=6.0.3",
    "tenacity>=9.1.4",
    "structlog>=25.5.0",
    "typing-extensions>=4.12.2",        # Backport type hints
    "numba>=0.60",                      # JIT and NJIT acceleration (used in perf paths)

    # Web/API (minimal dashboard/config on Pi)
    "httpx>=0.28.1",                    # For async HTTP client
    "fastapi>=0.135.1",
    "uvicorn>=0.42.0",
    "python-multipart>=0.0.22",

    # Auth & security
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "pyjwt[crypto]>=2.12.1",
    "qrcode[pil]>=8.2",
    "pyotp>=2.9.0",
    "zxcvbn>=4.5.0",

    # MQTT (async communication with Assistant)
    "asyncio-mqtt>=0.16.2",             # ASYNC MQTT client for communications
    "orjson>=3.11.7",                   # Fast JSON for MQTT payloads

    # Database
    "aiosqlite>=0.22.1",                # Async SQLite for FAISS/incident DB
    "faiss-cpu>=1.13.2",                # Vector DB for RAG/incident retrieval

    # ML Basics
    "torch-geometric>=2.7.0",           # GNN support (shared topology reasoning)
    "numpy>=2.4.3",                     # Array basics
    "torch>=2.10.0",                    # PyTorch base (for GNN/RL shared)
    "networkx>=3.6.1",                  # Graph utils (topology)
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.10.0",
    "mypy>=1.13",
    "ruff>=0.7",
    "pre-commit>=4.0",
    "types-pyyaml",
    "types-setuptools",
]

[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

[tool.setuptools.package-data]
"*" = ["*.pyi", "*.typed"]  # Include type stubs if you add them later